renamed legacy `UserService` imports to fix build errors
