+++ pages/account/edit.jsx
@@ -3,7 +3,7 @@ import Cookies from "cookies";
 import React from "react";
 import { useState, useEffect } from "react";
 import EditUserForm from "../../components/account/EditUserForm.component";
-import { getUser } from "../../services/UserService";
+import { getUser } from "../../services/AzureUserService";
 
 const EditUserPage = (props) => {
   const userId = props.userId;
